fix: darglint issues
we keep ignoring darglint checks for packages/valory/protocols on CI
because we cannot change some Python modules of the automatically
generated protocol modules (the other check with scripts/generate_all_protocols
would fail). However, the checks for that directory has been reintroduced
in the Makefile, so to help development.