[refactor]: change estimation payload
Change `EstimationPayload` to accept only a single estimation, because it is impossible to accept a list. Currently, a `steps_forward != 1` in the `skill.yaml` will fail for that reason. In the future, if more `steps_forward` are needed, this can be extended by saving the list of the latest observations (or even of the last `n` observations) to a file and passing the file's hash to the payload.