fix: decode_varint now shifts result back
in _TendermintABCISerializer, encode_varint left-shifts the input
by one, but decode_varint did not right-shift back the return value,
requiring the client code to shift the output manually.

This commit moves the right-shifts into the function decode_varint,
right before it is returned to the caller.

Moreover, this commit adds a test using Hypothesis to test that encoding
and decoding works correctly for integers sampled by the testing library.