fix: abci connection: update laddr from 0.0.0.0 to localhost
As pointed out in:

    https://github.com/valory-xyz/oracle-off-chain-logic-audit/blob/cb13cdc8ce56fad8bf7e96f293584c12989360c0/abci_connection_protocol_tendermint/README.md

It is better to bind the listening address to a specific address,
e.g. localhost, instead of 0.0.0.0 whenever possible.

This commit makes "localhost" the default listening address both for the ABCI app server
and for the RPC listening address. The reason is that very often the
interaction between:

- the Tendermint node and the ABCI App server in the case of the proxy_app address, and
- the client and the Tendermint rpc-server

occurs within the same machine.