feat: reset on disagreement about submitted tx
If a `NEGATIVE` result has been retrieved from the voting round, a tx the agents disagree on has been included.
This issue cannot occur if the keeper has done the job correctly and of course if the majority of the agents is not corrupted. This issue reduces to a case in which we have to re-select a keeper.