feat: try to send with a higher gas fee
The only one who can cancel the transaction is the sender, but actually cancelling is only possible by sending a higher priority tx with the same nonce. Here, the only two things the sender can do are:
- resend the existing tx with more gas fee (same as VALIDATE_TIMEOUT)
- send an arbitrary other tx with the same nonce.

The problem with the latter is that it's unnecessarily wasteful. So the only option is that we do the same as in the case of `VALIDATE_TIMEOUT` and simply try to send with a higher gas fee.