fix: check if we should choose a new keeper after `NONE`
The transition to FinalizationRound would not consider the fact that we are reusing the same keeper. Only `CollectSignatureRound` (and of course the select keeper b rounds) needs to transition immediately to `FinalizationRound`. In any other case, we can always first transition to `SelectKeeperTransactionSubmissionRoundB` before `FinalizationRound`, in order to select the appropriate keeper first.