feat: create parameters for various timeouts
Create parameters for the validation, finalization and tx history check timeouts.

Since we are now using a backoff factor it does not make sense anymore to use the `retry_timeout * retry_attempts + margin` in order to calculate the validation and tx check timeouts, because the result will always be smaller than the max backoff. Therefore, we need to use a separate parameter in order to create the following condition:
```python
min(retry_timeout ** n_retries, max_backoff)
```

where `max_backoff` is any of the new timeouts (`validate_timeout`, `history_check_timeout`).

Also added the `finalize_timeout` since there was a `TOFIX` comment for that.