refactor: update initial height's calculation
Initial height needs to cover the time between transition and reset. Since every second roughly one block is created we use: observation_interval / 2 + height_on_transition + offset. `observation_interval / 2` is the time required from the moment the agent enters the reset round until the time it starts to reset. Offset is used in order to account for shifts among the agents' sync and for the small amount of time required in order to actually reset the tendermint node. The bigger the possibility of lag among the agents, the larger the offset should be.