refactor: move `safe_contract_address` to the base
Move `safe_contract_address` property to the `BaseSynchronizedData`. It is still overriden in `packages/valory/skills/termination_abci/rounds.py::SynchronizedData`, because it is necessary for the `termination_abci` to function properly. In essence, we cannot assume that the registration round is over when the `safe_contract_address` is called by the `TerminationRound` since it runs concurrently, from the beginning of the app.