fix: `headers` and `parameters` in `ApiSpecs`
The issue was that the type was defined as `List[Tuple[str, str]]`. However, a `tuple` will never occur when loading from a `yaml`. Moreover, conceptually it does not make sense to use a list for these configurations. Therefore, they were both converted to a mapping, which is more correct.