refactor: update based on ledger api protocol refactoring
Update flashbots support that was introduced on 7b578fdca26fc86232186f00bdf76f62c0f19b83, based on the open-aea refactoring on https://github.com/valory-xyz/open-aea/pull/594.

Depends on valory-xyz/open-aea#592 which is merged and released at the moment of performing this commit, but the `open-autonomy` is not bumped to this version yet.

[no ci]