test: reveal the issue with `is_termination_set` property
The `is_termination_set` property will give its value to `include_termination_rounds` in the introduced test.

The property is used at the `_MetaRoundBehaviour` metaclass. As a result, it always gets interpreted as `True` in the `get_all_round_classes` as it is never evaluated, but results in the `property` object. Therefore, `bool(property_object) is True`.