fix: check sync after gentle resetting
That way, we sleep for a max time of `tendermint_check_sleep_delay * max_retries` and we also have the chance to reset the `gentle_reset_attempted` flag.

Otherwise, we could end up in a situation in which we successfully reset and then do not check sync, lose connection again, and attempt a hard reset at once because of the flag not being reset.