refactor: slashing is a separate app to avoid dependencies
If the slashing was tied to the absract round, it would create a dependency of the core with the tx settlement skill and the gnosis safe contract.

Instead, we make slashing optional, by extracting it to an abci app. This means that we cannot directly associate the slashing app with the core, and every app dev will have to add a few lines of code in order to add the chaining (the same way it is done for the termination).